+++ CHANGELOG.md
@@ -3,11 +3,17 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [3.6.8] - 2025-4-21
+
+### Updated
+
+- Bumped support for Keycloak libraries to 26.2.0
+
 ## [3.6.7] - 2025-4-07
 
 ### Updated
 
-- Bumped upport for Keycloak libraries to 26.1.4
+- Bumped support for Keycloak libraries to 26.1.4
 
 ## [3.6.6] - 2025-3-28